nanopore-basecaller/requirements.txt
numpy
pandas
torch
scikit-learn
matplotlib
seaborn
jupyter
tqdm
pyyaml
fast5
h5py